1. Add radius-dependent thinning to the pellet model + docs 2. Update the conveyor example to test ShapePack 3. Make conveyor compatible with ShapePack including z-trimming and such (to be thoroughly tested!) 4. Particle generators automatically say whether they are spheres-only or not 5. Enhance ShapePack API to be mostly compatible with SpherePack to ease transition

+++ pkg/dem/Particle.cpp
@@ -75,7 +75,7 @@ int Particle::countRealContacts() const{
 void Shape::setFromRaw_helper_checkRaw_makeNodes(const vector<Real>& raw, size_t numRaw){
 	if(raw.size()!=numRaw) throw std::runtime_error("Error setting "+pyStr()+" from raw data: "+to_string(numRaw)+" numbers expected, "+to_string(raw.size())+" given.");
 	// add/remove nodes as necessary
-	while(nodes.size()<numNodes()) nodes.push_back(make_shared<Node>());
+	while((int)nodes.size()<numNodes()) nodes.push_back(make_shared<Node>());
 	nodes.resize(numNodes());
 }
